yzy_investigation/projects/discord_manager/requirements.txt
# Discord Manager Project Requirements
discord.py>=2.3.2
python-dotenv>=1.0.0
aiohttp>=3.9.1
aiofiles>=23.2.1
python-dateutil>=2.8.2
openai>=1.12.0
tqdm>=4.66.1 